Separate draw module from utils
